+++ src/main/java/donmani/donmani_server/feedback/provider/FeedbackTemplate.java
@@ -1,4 +1,4 @@
-package donmani.donmani_server.feedback.entity;
+package donmani.donmani_server.feedback.provider;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.ToString;
