+++ src/main/java/donmani/donmani_server/feedback/service/FeedbackService.java
@@ -14,8 +14,8 @@ import donmani.donmani_server.expense.dto.ExpenseRequestDTO;
 import donmani.donmani_server.expense.entity.Expense;
 import donmani.donmani_server.expense.service.ExpenseService;
 import donmani.donmani_server.feedback.dto.FeedbackOpenResponseDTO;
-import donmani.donmani_server.feedback.entity.FeedbackTemplate;
-import donmani.donmani_server.feedback.entity.FeedbackTemplateProvider;
+import donmani.donmani_server.feedback.provider.FeedbackTemplate;
+import donmani.donmani_server.feedback.provider.FeedbackTemplateProvider;
 import donmani.donmani_server.user.entity.User;
 import donmani.donmani_server.user.service.UserService;
 import donmani.donmani_server.feedback.entity.Feedback;
